lib/analysis/taint: move tests for `State`
